refactor(verification): relocate validator utility
