+++ src/main/java/com/qibergames/futura/verification/Validator.java
@@ -1,4 +1,4 @@
-package com.qibergames.futura.util;
+package com.qibergames.futura.verification;
 
 import lombok.experimental.UtilityClass;
 import org.jetbrains.annotations.NotNull;
